+++ packages/orq-rc/src/models/operations/createpromptversion.ts
@@ -358,7 +358,6 @@ export type CreatePromptVersionMetadata = {
 };
 
 export type CreatePromptVersionRequestBody = {
-  id: string;
   displayName: string;
   description?: string | null | undefined;
   promptConfig: CreatePromptVersionPromptConfig;
@@ -371,7 +370,7 @@ export type CreatePromptVersionRequest = {
   /**
    * Prompt ID
    */
-  promptId: string;
+  id: string;
   requestBody?: CreatePromptVersionRequestBody | undefined;
 };
 
@@ -2051,7 +2050,6 @@ export const CreatePromptVersionRequestBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string(),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   prompt_config: z.lazy(() => CreatePromptVersionPromptConfig$inboundSchema),
@@ -2060,7 +2058,6 @@ export const CreatePromptVersionRequestBody$inboundSchema: z.ZodType<
   timestamp: z.string(),
 }).transform((v) => {
   return remap$(v, {
-    "_id": "id",
     "display_name": "displayName",
     "prompt_config": "promptConfig",
   });
@@ -2068,7 +2065,6 @@ export const CreatePromptVersionRequestBody$inboundSchema: z.ZodType<
 
 /** @internal */
 export type CreatePromptVersionRequestBody$Outbound = {
-  _id: string;
   display_name: string;
   description?: string | null | undefined;
   prompt_config: CreatePromptVersionPromptConfig$Outbound;
@@ -2083,7 +2079,6 @@ export const CreatePromptVersionRequestBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreatePromptVersionRequestBody
 > = z.object({
-  id: z.string(),
   displayName: z.string(),
   description: z.nullable(z.string()).optional(),
   promptConfig: z.lazy(() => CreatePromptVersionPromptConfig$outboundSchema),
@@ -2092,7 +2087,6 @@ export const CreatePromptVersionRequestBody$outboundSchema: z.ZodType<
   timestamp: z.string(),
 }).transform((v) => {
   return remap$(v, {
-    id: "_id",
     displayName: "display_name",
     promptConfig: "prompt_config",
   });
@@ -2137,19 +2131,18 @@ export const CreatePromptVersionRequest$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  prompt_id: z.string(),
+  id: z.string(),
   RequestBody: z.lazy(() => CreatePromptVersionRequestBody$inboundSchema)
     .optional(),
 }).transform((v) => {
   return remap$(v, {
-    "prompt_id": "promptId",
     "RequestBody": "requestBody",
   });
 });
 
 /** @internal */
 export type CreatePromptVersionRequest$Outbound = {
-  prompt_id: string;
+  id: string;
   RequestBody?: CreatePromptVersionRequestBody$Outbound | undefined;
 };
 
@@ -2159,12 +2152,11 @@ export const CreatePromptVersionRequest$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreatePromptVersionRequest
 > = z.object({
-  promptId: z.string(),
+  id: z.string(),
   requestBody: z.lazy(() => CreatePromptVersionRequestBody$outboundSchema)
     .optional(),
 }).transform((v) => {
   return remap$(v, {
-    promptId: "prompt_id",
     requestBody: "RequestBody",
   });
 });